update destory  purchase

+++ app/Http/Controllers/purchase/purchaseController.php
@@ -321,16 +321,18 @@ class purchaseController extends Controller
                 foreach ($purchaseDetails->get() as $pd) {
                     $csQuery = CurrentStockBalance::where('transaction_type', 'purchase')->where('transaction_detail_id', $pd->id)->first();
 
-                    if ($csQuery->current_quantity < $csQuery->ref_uom_quantity) {
-                        throw new Exception("Can't Delete This Purchase Transactions. Because some stocks are already out from this transactions.");
-                    } else {
-                        $csQuery->delete();
-                        stock_history::where('transaction_type', 'purchase')->where('transaction_details_id', $pd->id)->delete();
-                        $purchaseDetails->update([
-                            'is_delete' => 1,
-                            'deleted_by' => Auth::user()->id,
-                            'deleted_at' => now()
-                        ]);
+                    if ($csQuery) {
+                        if ($csQuery->current_quantity < $csQuery->ref_uom_quantity) {
+                            throw new Exception("Can't Delete This Purchase Transactions. Because some stocks are already out from this transactions.");
+                        } else {
+                            $csQuery->delete();
+                            stock_history::where('transaction_type', 'purchase')->where('transaction_details_id', $pd->id)->delete();
+                            $purchaseDetails->update([
+                                'is_delete' => 1,
+                                'deleted_by' => Auth::user()->id,
+                                'deleted_at' => now()
+                            ]);
+                        }
                     }
                 }
             }
